Create Original
The original functional source code of the Rett Project. This provides the basic functionality of the project; dividing the FastQ file into Genes, scanning those gene for hits, and then adding those hits to a list so that the total number of FOXO3 binding motif containing genes in the data-set can be calculated.